requirements.txt
# =============================================================================
# FORESIGHT — Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
# PostgreSQL (async)
asyncpg==0.29.0
psycopg2-binary==2.9.9

# MongoDB
motor==3.3.2
pymongo==4.6.1

# SQLAlchemy (for Airflow compatibility)
sqlalchemy==2.0.25

# -----------------------------------------------------------------------------
# Message Queue
# -----------------------------------------------------------------------------
kafka-python==2.0.2
confluent-kafka==2.3.0

# -----------------------------------------------------------------------------
# Apache Spark
# -----------------------------------------------------------------------------
pyspark==3.5.0

# -----------------------------------------------------------------------------
# Apache Airflow
# -----------------------------------------------------------------------------
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-mongo==3.5.0

# -----------------------------------------------------------------------------
# Machine Learning
# -----------------------------------------------------------------------------
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# MLflow
mlflow==2.10.0

# -----------------------------------------------------------------------------
# Object Storage (MinIO/S3)
# -----------------------------------------------------------------------------
boto3==1.34.0
botocore==1.34.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pyarrow==15.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
httpx==0.26.0
requests==2.31.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.0
flake8==7.0.0
mypy==1.8.0

# -----------------------------------------------------------------------------
# Type Stubs
# -----------------------------------------------------------------------------
types-requests==2.31.0
types-python-jose==3.3.4
types-passlib==1.7.7
